rgw:cleanup/refactor json and xml encoders and decoders
move the encoder and decoder methods into their associated class
files to eliminate undefined references to the class vtable

https://tracker.ceph.com/issues/53596
