+++ src/rgw/rgw_acl_s3.cc
@@ -631,3 +631,13 @@ XMLObj *RGWACLXMLParser_S3::alloc_obj(const char *el)
   return obj;
 }
 
+ACLGroupTypeEnum ACLGrant_S3::uri_to_group(string& uri)
+{
+  if (uri.compare(rgw_uri_all_users) == 0)
+    return ACL_GROUP_ALL_USERS;
+  else if (uri.compare(rgw_uri_auth_users) == 0)
+    return ACL_GROUP_AUTHENTICATED_USERS;
+
+  return ACL_GROUP_NONE;
+}
+
