+++ src/rgw/rgw_compression.cc
@@ -208,3 +208,21 @@ int RGWGetObj_Decompress::fixup_range(off_t& ofs, off_t& end)
 
   return next->fixup_range(ofs, end);
 }
+
+void compression_block::dump(Formatter *f) const
+{
+  f->dump_unsigned("old_ofs", old_ofs);
+  f->dump_unsigned("new_ofs", new_ofs);
+  f->dump_unsigned("len", len);
+}
+
+void RGWCompressionInfo::dump(Formatter *f) const
+{
+  f->dump_string("compression_type", compression_type);
+  f->dump_unsigned("orig_size", orig_size);
+  if (compressor_message) {
+    f->dump_int("compressor_message", *compressor_message);
+  }
+  ::encode_json("blocks", blocks, f);
+}
+
